Add Material Design Library as kick off
