+++ tools/Profiler/NanoProfiler/App.xaml
@@ -1,16 +1,20 @@
 ﻿<Application
     x:Class="nanoFramework.Tools.NanoProfiler.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-     
-    >
+    xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+    StartupUri="Views/ShellWindow.xaml"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" >
     <!--StartupUri="ProfilerLauncher.xaml"-->
     <Application.Resources>
         <ResourceDictionary> 
             <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="pack://application:,,,/NanoProfiler.Themes;component/Assets/NanoProfiler.xaml" />
-                <ResourceDictionary Source="pack://application:,,,/NanoProfiler.Themes;component/Assets/NanoProfiler.Light.xaml" />
+                <materialDesign:BundledTheme BaseTheme="Light"
+                                             ColorAdjustment="{materialDesign:ColorAdjustment}"
+                                             PrimaryColor="BlueGrey"
+                                             SecondaryColor="Lime" />
+                <ResourceDictionary Source="pack://application:,,,/MaterialDesignThemes.Wpf;component/Themes/MaterialDesign3.Defaults.xaml" />
+               
             </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
+        </ResourceDictionary>         
     </Application.Resources>
 </Application>
