+++ tools/Profiler/NanoProfiler/Settings/SettingsView.xaml
@@ -1,0 +1,44 @@
+﻿<UserControl x:Class="nanoFramework.Tools.NanoProfiler.Settings.SettingsView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:nanoFramework.Tools.NanoProfiler.Settings"
+             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+
+    <!--<materialDesign:Card Background="{DynamicResource MaterialDesignPaper}"
+                         UniformCornerRadius="15"
+                         Margin="2"
+                         materialDesign:ElevationAssist.Elevation="Dp3">-->
+    <Grid Background="{Binding BackgroundProfileLauncher}">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="290" />
+                <ColumnDefinition Width="*" />
+            </Grid.ColumnDefinitions>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto"></RowDefinition>
+                <RowDefinition Height="Auto"></RowDefinition>
+                <RowDefinition Height="*"></RowDefinition>
+                <RowDefinition Height="Auto"></RowDefinition>
+            </Grid.RowDefinitions>
+
+            <TextBlock Margin="10"
+                       Grid.Row="0"                        
+                       Text="Theme:"
+                       Style="{DynamicResource MaterialDesignHeadline5TextBlock}"></TextBlock>
+            <StackPanel Margin="10" Grid.Row="1" Orientation="Horizontal">
+                <TextBlock VerticalAlignment="Center"
+                           Text="Light" 
+                           FontSize="18" />  
+                <ToggleButton Margin="8,0,11,0"   
+                              IsChecked="{Binding IsDarkTheme}" />
+                <TextBlock VerticalAlignment="Center"
+                       FontSize="18" 
+                       Text="Dark" />
+            </StackPanel>   
+
+        </Grid>
+    <!--</materialDesign:Card>-->
+</UserControl>
