--- tools/Profiler/NanoProfiler/Views/ProfilerLauncherView.xaml
@@ -1,98 +1,0 @@
-﻿<Window x:Class="nanoFramework.Tools.NanoProfiler.Views.ProfilerLauncherView"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:nanoFramework.Tools.NanoProfiler.Views"
-        xmlns:viewModel="clr-namespace:nanoFramework.Tools.NanoProfiler.ViewModels"
-        mc:Ignorable="d"
-        Title=".NET nanoFramework Profiler Launcher" Height="579" VerticalContentAlignment="Stretch" HorizontalContentAlignment="Stretch"
-        Unloaded="Window_Unloaded" Loaded="Window_Loaded" Icon="/Views/nanoframework.ico" Left="0" Top="0" Width="834">
-    <Window.DataContext>
-        <viewModel:ProfilerLauncherViewModel/>
-    </Window.DataContext>
-    <Grid Background="{Binding BackgroundProfileLauncher}" HorizontalAlignment="Stretch" VerticalAlignment="Stretch">
-
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-        </Grid.RowDefinitions>
-
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="Auto"/>
-            <ColumnDefinition Width="Auto"/>
-            <ColumnDefinition Width="Auto"/>
-        </Grid.ColumnDefinitions>
-
-        <GroupBox Header="Device" Grid.Row="0" Grid.Column="0" Width="300" Margin="20,20,0,0" >
-            <Grid>
-                <Label Content="nanoDevice to connect to:" HorizontalAlignment="Left" Margin="10,10,0,0" VerticalAlignment="Top"/>
-                <TextBox HorizontalAlignment="Left" Margin="18,41,0,0" TextWrapping="Wrap" Text="{Binding ComPortName}" VerticalAlignment="Top" Width="120"/>
-                <Button x:Name="ConnectButton" IsEnabled="{Binding ConnectButtonEnabled}"
-                Content="{Binding ConnectButtonContent}" HorizontalAlignment="Left" Margin="18,69,0,0" VerticalAlignment="Top" Width="120" 
-                Command="{Binding ConnectClickedCommand}"/>
-            </Grid>
-        </GroupBox>
-
-        <GroupBox Header="Profiling Options" Grid.Row="0" Grid.Column="1" Width="300" Margin="20,20,0,0" >
-            <Grid >
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="*"/>
-                </Grid.RowDefinitions>
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="284*"/>
-                </Grid.ColumnDefinitions>
-                <CheckBox Grid.Column="0" Grid.Row="0" Content="Force Reboot on Connect" HorizontalAlignment="Left" Margin="10,10,0,0" 
-                          Name="checkReboot"  Height="23" IsChecked="{Binding RebootChecked}"/>
-                <CheckBox Grid.Column="0" Grid.Row="1" Content="Profile Function Calls" HorizontalAlignment="Left" Margin="10,0,0,0" 
-                          Name="checkCalls" IsChecked="{Binding CallsChecked}" Height="23"/>
-                <CheckBox Grid.Column="0" Grid.Row="2" Content="Profile Object Allocations" HorizontalAlignment="Left" Margin="10,0,0,0" 
-                          Name="checkAllocations" IsChecked="{Binding AllocationsChecked}" Height="23"/>
-                <CheckBox Grid.Column="0" Grid.Row="3" Content="User Heap Absolute Address" HorizontalAlignment="Left" Margin="10,0,0,0" 
-                          Name="HeapAbsoluteAddress" Height="23" IsChecked="{Binding HeapAbsoluteAddressChecked}"/>
-            </Grid>
-        </GroupBox>
-
-        <GroupBox Header="Output"  Grid.Row="1" Grid.ColumnSpan="2" Grid.Column="0" Margin="20,20,0,0" >
-            <Grid>
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                </Grid.RowDefinitions>
-                
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="77*"/>
-                    <ColumnDefinition Width="44*"/>
-                    <ColumnDefinition Width="441*"/>
-                </Grid.ColumnDefinitions>
-                
-                <Label Content="File:" Margin="10,10,35,-3" Grid.Row="0"/>
-                <TextBox HorizontalAlignment="Left" Margin="47,15,0,0" TextWrapping="Wrap" IsReadOnly="True" VerticalAlignment="Top" Width="399" Grid.ColumnSpan="3" Text="{Binding OutputFileName}"  Grid.Row="0"/>
-                <Button Content="Browse..." Command="{Binding BrowseLogFileCommand}" Grid.Column="2" Margin="354,15,10,6"  Grid.Row="0"></Button>
-
-                <CheckBox Content="Trace profiler events" Grid.Row="1" Margin="20,20,0,20" IsChecked="{Binding TraceProfilesEvents}" />
-            </Grid>
-        </GroupBox>
-
-        <Grid Grid.Row="2" Grid.ColumnSpan="2" Grid.Column="0" Margin="20,20,0,0" VerticalAlignment="Top" Grid.RowSpan="2">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto"/>
-                <RowDefinition Height="Auto"/>
-                <RowDefinition Height="*"/>
-            </Grid.RowDefinitions>
-            <TextBox x:Name="textLog" Grid.Row="0"
-                 TextWrapping="Wrap" VerticalAlignment="Top" HorizontalAlignment="Stretch" Height="180" VerticalScrollBarVisibility="Auto"/>
-            <Button Content="Clear" Grid.Row="1" HorizontalAlignment="Left" VerticalAlignment="Top" Width="91" 
-                Command="{Binding ClearClickedCommand}" Margin="0,20,0,20"/>
-        </Grid>
-        
-        <Button x:Name="OpenLogFile" Content="Open Log File" HorizontalAlignment="Center" Grid.Column="2" Grid.Row="0"
-            VerticalAlignment="Center" Width="103" Command="{Binding OpenLogFileCommand}" Margin="40,0,40,0" />
-
-    </Grid>
-</Window>
